big ol gui refactor
