+++ src/main/java/com/palehorsestudios/alone/gui/GameManager.java
@@ -1,0 +1,16 @@
+package com.palehorsestudios.alone.gui;
+
+import com.palehorsestudios.alone.player.Player;
+
+public class GameManager {
+    private Player player;
+
+
+    public void setPlayer(Player player) {
+        this.player = player;
+    }
+
+    public Player getPlayer() {
+        return player;
+    }
+}
