--- src/main/java/com/palehorsestudios/alone/gui/IntroController.java
@@ -1,24 +1,0 @@
-package com.palehorsestudios.alone.gui;
-
-import javafx.fxml.FXML;
-import javafx.scene.control.Button;
-import javafx.scene.control.TextArea;
-
-public class IntroController {
-
-    @FXML
-    private Button startGame;
-    @FXML
-    private TextArea intro;
-
-    public IntroController() {
-    }
-
-    public Button getStartGame() {
-        return startGame;
-    }
-
-    public TextArea getIntro() {
-        return intro;
-    }
-}
