+++ src/main/java/com/palehorsestudios/alone/gui/controller/ItemSelectionWindowController.java
@@ -1,18 +1,30 @@
-package com.palehorsestudios.alone.gui;
+package com.palehorsestudios.alone.gui.controller;
 
 import com.palehorsestudios.alone.Items.Item;
 import com.palehorsestudios.alone.Items.ItemFactory;
+import com.palehorsestudios.alone.gui.GameManager;
+import com.palehorsestudios.alone.gui.ViewFactory;
+import com.palehorsestudios.alone.player.Player;
+import javafx.animation.KeyFrame;
+import javafx.animation.Timeline;
 import javafx.beans.value.ChangeListener;
 import javafx.beans.value.ObservableValue;
+import javafx.event.Event;
+import javafx.event.EventHandler;
 import javafx.fxml.FXML;
+import javafx.fxml.Initializable;
 import javafx.scene.control.Button;
 import javafx.scene.control.CheckBox;
 import javafx.scene.control.Label;
 import javafx.scene.layout.GridPane;
+import javafx.stage.Stage;
 
+import java.net.URL;
 import java.util.*;
 
-public class ItemSelectionController {
+import static javafx.util.Duration.seconds;
+
+public class ItemSelectionWindowController extends BaseController implements Initializable {
     @FXML
     private GridPane paneSelected;
     @FXML
@@ -82,10 +94,26 @@ public class ItemSelectionController {
     @FXML
     private CheckBox journalandpen;
 
-    private static int count = 0;
+    // Private Variables
+    private static final int COUNT_DOWN = 30;
+    private int count = 0;
     private List<Item> initItems = new ArrayList<>();
 
-    public void selectItems() {
+    public ItemSelectionWindowController(GameManager gameManager, ViewFactory viewFactory, String fxmlName) {
+        super(gameManager, viewFactory, fxmlName);
+    }
+
+    @FXML
+    public void playGameButton() {
+        System.out.println("Play button pressed");
+        // Player now managed in gameManager
+        gameManager.setPlayer(new Player(initItems));
+        viewFactory.showGameWindow();
+        Stage stage = (Stage) countdown.getScene().getWindow();
+        viewFactory.closeStage(stage);
+    }
+
+    private void selectItems() {
         Map<CheckBox, Item> inventory = new HashMap<>();
         inventory.put(fishingLine, ItemFactory.getNewInstance("Fishing Line"));
         inventory.put(fishingHooks, ItemFactory.getNewInstance("Fishing Hooks"));
@@ -158,5 +186,37 @@ public class ItemSelectionController {
 
     public List<Item> getInitItems() {
         return initItems;
+    }
+
+    @FXML
+    @Override
+    public void initialize(URL location, ResourceBundle resources) {
+        startTimer();
+        selectItems();
+    }
+
+    private void startTimer() {
+        // start the count down timer
+        Integer[] timeSeconds = {COUNT_DOWN};
+        Timeline timeline = new Timeline();
+        timeline.setCycleCount(Timeline.INDEFINITE);
+        timeline
+                .getKeyFrames()
+                .add(
+                        new KeyFrame(
+                                seconds(1),
+                                new EventHandler() {
+                                    @Override
+                                    public void handle(Event event) {
+                                        timeSeconds[0]--;
+                                        // update timerLabel
+                                        getCountdown().setText(timeSeconds[0].toString());
+                                        if (timeSeconds[0] <= 0) {
+                                            getPaneSelected().setDisable(true);
+                                            timeline.stop();
+                                        }
+                                    }
+                                }));
+        timeline.playFromStart();
     }
 }
