--- src/main/java/com/palehorsestudios/alone/gui/intro.fxml
@@ -1,95 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.*?>
-<?import javafx.scene.effect.InnerShadow?>
-<?import javafx.scene.effect.SepiaTone?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.text.Font?>
-<?import java.net.URL?>
-<VBox prefHeight="676.0" prefWidth="903.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1">
-    <children>
-        <MenuBar prefHeight="0.0" prefWidth="640.0" VBox.vgrow="NEVER">
-            <menus>
-                <Menu mnemonicParsing="false" text="File">
-                    <items>
-                        <MenuItem mnemonicParsing="false" text="New"/>
-                        <MenuItem mnemonicParsing="false" text="Open…"/>
-                        <Menu mnemonicParsing="false" text="Open Recent"/>
-                        <SeparatorMenuItem mnemonicParsing="false"/>
-                        <MenuItem mnemonicParsing="false" text="Close"/>
-                        <MenuItem mnemonicParsing="false" text="Save"/>
-                        <MenuItem mnemonicParsing="false" text="Save As…"/>
-                        <MenuItem mnemonicParsing="false" text="Revert"/>
-                        <SeparatorMenuItem mnemonicParsing="false"/>
-                        <MenuItem mnemonicParsing="false" text="Preferences…"/>
-                        <SeparatorMenuItem mnemonicParsing="false"/>
-                        <MenuItem mnemonicParsing="false" text="Quit"/>
-                    </items>
-                </Menu>
-                <Menu mnemonicParsing="false" text="Edit">
-                    <items>
-                        <MenuItem mnemonicParsing="false" text="Undo"/>
-                        <MenuItem mnemonicParsing="false" text="Redo"/>
-                        <SeparatorMenuItem mnemonicParsing="false"/>
-                        <MenuItem mnemonicParsing="false" text="Cut"/>
-                        <MenuItem mnemonicParsing="false" text="Copy"/>
-                        <MenuItem mnemonicParsing="false" text="Paste"/>
-                        <MenuItem mnemonicParsing="false" text="Delete"/>
-                        <SeparatorMenuItem mnemonicParsing="false"/>
-                        <MenuItem mnemonicParsing="false" text="Select All"/>
-                        <MenuItem mnemonicParsing="false" text="Unselect All"/>
-                    </items>
-                </Menu>
-                <Menu mnemonicParsing="false" text="Help">
-                    <items>
-                        <MenuItem mnemonicParsing="false" text="About MyHelloApp"/>
-                    </items>
-                </Menu>
-            </menus>
-        </MenuBar>
-        <AnchorPane maxHeight="-1.0" maxWidth="-1.0" prefHeight="-1.0" prefWidth="-1.0" VBox.vgrow="ALWAYS">
-            <children>
-                <AnchorPane layoutX="57.0" layoutY="33.0" prefHeight="371.0" prefWidth="640.0"
-                            AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0"
-                            AnchorPane.topAnchor="0.0">
-                    <children>
-                        <TextArea fx:id="intro" editable="false" layoutX="8.0" layoutY="38.0" prefHeight="565.0"
-                                  prefWidth="903.0" wrapText="true" AnchorPane.bottomAnchor="44.0"
-                                  AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="38.0">
-                            <effect>
-                                <SepiaTone>
-                                    <input>
-                                        <InnerShadow/>
-                                    </input>
-                                </SepiaTone>
-                            </effect>
-                        </TextArea>
-                        <Button fx:id="startGame" layoutX="298.0" layoutY="344.0" mnemonicParsing="false"
-                                prefHeight="45.0" prefWidth="903.0" text="Start Game" AnchorPane.bottomAnchor="0.0"
-                                AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0">
-                            <font>
-                                <Font size="24.0"/>
-                            </font>
-                            <effect>
-                                <InnerShadow/>
-                            </effect>
-                        </Button>
-                        <TextField alignment="CENTER" editable="false" layoutX="102.0" layoutY="14.0"
-                                   promptText="Game Introduction" text="Game Introduction" AnchorPane.leftAnchor="0.0"
-                                   AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
-                            <font>
-                                <Font size="19.0"/>
-                            </font>
-                            <effect>
-                                <InnerShadow/>
-                            </effect>
-                        </TextField>
-                    </children>
-                </AnchorPane>
-            </children>
-        </AnchorPane>
-    </children>
-    <stylesheets>
-        <URL value="@BasicApplication.css"/>
-    </stylesheets>
-</VBox>
