+++ src/main/java/com/palehorsestudios/alone/gui/view/css/BasicApplication.css
