+++ src/test/java/com/palehorsestudios/alone/MainTest.java
