+++ src/test/java/com/palehorsestudios/alone/ShelterTest.java
